// src/app/page/master/map-produk-satuanstandar/map-produk-satuanstandar.interface.ts
import {EventEmitter} from '@angular/core';

export interface MapProdukSatuanstandar {
	kdProduk?: string;
	kdEvent?: number;

	kode?: any;
	id?: any;
	kdProfile?: string;
	version?: string;
	reportDisplay?: string;
	kodeExternal?: string;
	namaExternal?: string;
	statusEnabled?: boolean;
	noRec?: string;
	label?: any;
	value?: any;
}